end of basket

+++ Core/Entities/Product.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace Core.Entities
@@ -6,7 +7,7 @@ namespace Core.Entities
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public ICollection<Photo> Photos { get; set; }
         public ProductType ProductType { get; set; }
         public int ProductTypeId { get; set; }
